PortFolio용 프로젝트 생성
Scene 관련 코드 리팩토링 (GUI, Graphics 분리)

UI Component 형태로 제공 엔진 단으로 넘길지는 고민 필요

+++ Engine/04.Component/Camera.cpp
@@ -7,8 +7,8 @@ Matrix Camera::S_MatProjection = Matrix::Identity;
 
 Camera::Camera() : Super(ComponentType::Camera)
 {
-	_width = static_cast<float>(GAME->GetGameDesc().width);
-	_height = static_cast<float>(GAME->GetGameDesc().height);
+	_width = GAME->GetGameDesc().width;
+	_height = GAME->GetGameDesc().height;
 }
 
 Camera::~Camera()
